bulk message sending implemented

+++ client/src/pages/ValidateNumbers.jsx
@@ -1,6 +1,8 @@
 import {
   Box,
   Button,
+  Card,
+  CardBody,
   Heading,
   IconButton,
   Modal,
@@ -18,15 +20,47 @@ import {
 import React, { useEffect, useState } from "react";
 import { TbTrashXFilled } from "react-icons/tb";
 import { MdContactPhone, MdTask } from "react-icons/md";
+import vCardsJS from "vcards-js";
+import { Link, useNavigate } from "react-router-dom";
+
 export default function ValidateNumbers({ socket }) {
+  const navigate = useNavigate();
   const { isOpen, onOpen, onClose } = useDisclosure();
   const [allTasks, setAllTasks] = useState([]);
+  const [downloadUrl, setDownloadUrl] = useState("");
   const [commandToExecute, setCommandToExecute] = useState({
     command: "",
     message: "",
     taskId: "",
     confirm: false,
   });
+  function deleteTask(taskId) {
+    console.log({ commandToExecute });
+    socket.emit("delete_task", taskId);
+    setCommandToExecute({ command: "", message: "", confirm: false });
+    socket.emit("get_tasks", "");
+  }
+  function getTaskData(taskId) {
+    socket.emit("get_task_data", taskId);
+  }
+  useEffect(() => {
+    socket.on("task_data", (data) => {
+      var mainString = "";
+      const taskElements = data;
+      taskElements.forEach((contact) => {
+        var vCard = vCardsJS();
+        vCard.firstName = contact.name || "";
+        vCard.workPhone = contact.number || "";
+        vCard.email = contact.email || "";
+        mainString = mainString + vCard.getFormattedString();
+        console.log("mainString", mainString);
+      });
+      console.log("mainString final", mainString);
+      var vCardBlob = new Blob([mainString], { type: "text/vcard" });
+      var vCardURL = window.URL.createObjectURL(vCardBlob);
+      setDownloadUrl(vCardURL);
+    });
+  }, [socket]);
   useEffect(() => {
     socket.on("all_tasks", (data) => {
       const taskIds = Object.keys(data);
@@ -42,37 +76,43 @@ export default function ValidateNumbers({ socket }) {
     });
   }, [socket]);
   useEffect(() => {
-    socket.emit("get_all_contacts", { profilePicUrl: false });
+    socket.emit("get_all_contacts", { profilePicUrl: true });
     socket.emit("get_tasks", "");
   }, []);
-  function deleteTask(taskId) {
-    console.log({ commandToExecute });
-    socket.emit("delete_task", taskId);
-    setCommandToExecute({ command: "", message: "", confirm: false });
-    socket.emit("get_tasks", "");
-  }
+
   const ConfirmModal = ({ command, message }) => {
     return (
       <Modal isOpen={isOpen} onClose={onClose}>
         <ModalOverlay />
         <ModalContent>
-          <ModalHeader>{command}</ModalHeader>
+          <ModalHeader color="gray.600">{command}</ModalHeader>
           <ModalCloseButton />
           <ModalBody>{message}</ModalBody>
 
           <ModalFooter>
-            <Button colorScheme="blue" mr={3} onClick={onClose}>
-              Close
-            </Button>
             <Button
               colorScheme="whatsapp"
-              onClick={() => {
-                deleteTask(commandToExecute.taskId);
-                onClose();
-              }}
+              variant="outline"
+              mr={3}
+              onClick={onClose}
             >
-              Confirm
+              Close
             </Button>
+            {command == "Download" ? (
+              <Link to={downloadUrl} target="_blank">
+                <Button onClick={onClose}>{command}</Button>
+              </Link>
+            ) : (
+              <Button
+                colorScheme="whatsapp"
+                onClick={() => {
+                  if (command == "Delete") deleteTask(commandToExecute.taskId);
+                  onClose();
+                }}
+              >
+                {command}
+              </Button>
+            )}
           </ModalFooter>
         </ModalContent>
       </Modal>
@@ -96,76 +136,93 @@ export default function ValidateNumbers({ socket }) {
           minW="xl"
           padding="2"
           overflowY="auto"
+          alignItems="center"
         >
           {allTasks.length == 0 && <Text color="whiteAlpha.500">No Tasks</Text>}
           {allTasks.map((task) => {
             return (
-              <Stack
+              <Card
                 key={task.id}
                 bg="gray.800"
                 borderRadius="md"
                 color="whiteAlpha.800"
-                padding="2"
-                py="4"
-                direction="row"
-                justifyContent="space-between"
+                width="100%"
               >
-                <Stack>
-                  <Text fontWeight="bold" fontSize="lg" color="whatsapp.400">
-                    #
-                    <span
-                      style={{ color: "var(--chakra-colors-whiteAlpha-800)" }}
-                    >
-                      {task.id}
-                    </span>
-                  </Text>
-                  <Text color="whatsapp.700" fontSize="sm">
-                    {task.createdAt}
-                  </Text>
-                </Stack>
-                <Stack alignItems="flex-end">
-                  <Text color="whiteAlpha.400" fontSize="xs">
-                    <strong style={{ color: "whatsapp.400" }}>
-                      {task.length}
-                    </strong>{" "}
-                    {task.length > 1 ? "numbers" : "number"}
-                  </Text>
-                  <Stack direction="row">
-                    <Tooltip label="Delete task">
-                      <IconButton
-                        icon={<TbTrashXFilled />}
-                        colorScheme="blackAlpha"
-                        color="whiteAlpha.600"
-                        _hover={{ borderColor: "red.500", color: "red.500" }}
-                        onClick={() => {
-                          setCommandToExecute({
-                            ...commandToExecute,
-                            command: "Delete",
-                            taskId: task.id,
-                            message: `Are you sure you want to delete task #${task.id}?`,
-                          });
-                          onOpen();
-                        }}
-                      />
-                    </Tooltip>
-                    <Tooltip label="Download vCard">
-                      <IconButton
-                        icon={<MdContactPhone />}
-                        colorScheme="blackAlpha"
-                        color="whiteAlpha.600"
-                        _hover={{ color: "whatsapp.500" }}
-                      />
-                    </Tooltip>
-                    <Tooltip label="View results">
-                      <IconButton
-                        icon={<MdTask />}
-                        colorScheme="blackAlpha"
-                        color="whatsapp.500"
-                      />
-                    </Tooltip>
+                <CardBody
+                  display="flex"
+                  direction="row"
+                  justifyContent="space-between"
+                >
+                  <Stack>
+                    <Text fontWeight="bold" fontSize="lg" color="whatsapp.400">
+                      #
+                      <span
+                        style={{ color: "var(--chakra-colors-whiteAlpha-800)" }}
+                      >
+                        {task.id}
+                      </span>
+                    </Text>
+                    <Text color="whatsapp.700" fontSize="sm">
+                      {task.createdAt}
+                    </Text>
                   </Stack>
-                </Stack>
-              </Stack>
+                  <Stack alignItems="flex-end">
+                    <Text color="whiteAlpha.400" fontSize="xs">
+                      <strong style={{ color: "whatsapp.400" }}>
+                        {task.length}
+                      </strong>{" "}
+                      {task.length > 1 ? "numbers" : "number"}
+                    </Text>
+                    <Stack direction="row">
+                      <Tooltip label="Delete task">
+                        <IconButton
+                          icon={<TbTrashXFilled />}
+                          colorScheme="blackAlpha"
+                          color="whiteAlpha.600"
+                          _hover={{ borderColor: "red.500", color: "red.500" }}
+                          onClick={() => {
+                            setCommandToExecute({
+                              ...commandToExecute,
+                              command: "Delete",
+                              taskId: task.id,
+                              message: `Are you sure you want to delete task #${task.id}?`,
+                            });
+                            onOpen();
+                          }}
+                        />
+                      </Tooltip>
+                      <Tooltip label="Download vCard">
+                        <IconButton
+                          icon={<MdContactPhone />}
+                          colorScheme="blackAlpha"
+                          color="whiteAlpha.600"
+                          _hover={{ color: "whatsapp.500" }}
+                          onClick={() => {
+                            setCommandToExecute({
+                              ...commandToExecute,
+                              command: "Download",
+                              taskId: task.id,
+                              message: `Your vCard for task #${task.id} is ready.`,
+                            });
+                            getTaskData(task.id);
+                            onOpen();
+                          }}
+                        />
+                      </Tooltip>
+                      <Tooltip label="View results">
+                        <IconButton
+                          icon={<MdTask />}
+                          colorScheme="blackAlpha"
+                          color="whatsapp.500"
+                          onClick={() =>
+                            navigate(`/validateNumbers/${task.id}`)
+                          }
+                        />
+                      </Tooltip>
+                    </Stack>
+                  </Stack>
+                </CardBody>
+              </Card>
             );
           })}
         </Stack>
